// cloud/src/services/agentService/types.ts
export interface IncomingDevice {
  ip: string;
  serial: string | null;
  mac?: string | null;
  brand: string;
  model: string;
  name: string;
}